requirements.txt
Django==4.2.7
django-crispy-forms==2.0
crispy-tailwind==0.5.0
stripe==7.6.0
Pillow>=10.0.0
python-decouple==3.8
django-allauth==0.57.0
django-environ==0.11.2
django-taggit==4.0.0
django-star-ratings==0.9.2
django-ckeditor==6.7.0
django-cors-headers==4.3.1
whitenoise==6.6.0
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1 